fix: convert network/networkId using lookup table
Currently we use networkId for cacheKeys but this is not always set.
This means that we might end up using wrong cached asset.

To solve this we generate lookup table using chains data that will
minimize this issue (there can be still some cases of missing mapping,
but way less likely now).